Correct load to load cost caused by false dependency in vicuna
Consider the following example that causes a false dependency.
  r1 = sxn([r0])
  r0 = r0+0x4
  r2 = sxn([r0+4])
The 2nd load will be marked as "multiple dependency" to the 1st load.
Since it's not a REG_DEP_ANTI nor a REG_DEP_OUTPUT, it will be treated
as a REG_DEP_TRUE and cause unnecessary stalls when scheduling.
Fix the issue by checking if the dst of 1st load == base reg of 2nd load.

Bugzilla #24428